Update hinet sms struct to parse the response body.
